requirements.txt
soundfile
torchaudio
styletts2
fastapi
pydantic
uvicorn
pydub
python-magic